feat: implement account ownership validation and enhance account management operations

+++ db/query/account.sql
@@ -12,8 +12,9 @@ SELECT * FROM accounts WHERE id = $1 LIMIT 1;
 
 -- name: ListAccounts :many
 SELECT * FROM accounts
+WHERE owner = $1
 ORDER BY id
-LIMIT $1 OFFSET $2;
+LIMIT $2 OFFSET $3;
 
 -- name: UpdateAccount :one
 UPDATE accounts
@@ -21,8 +22,8 @@ SET balance = $2
 WHERE id = $1
 RETURNING *;
 
--- name: DeleteAccount :exec
-DELETE FROM accounts WHERE id = $1;
+-- name: DeleteAccount :one
+DELETE FROM accounts WHERE id = $1 AND owner = $2 RETURNING *;
 
 -- name: GetAccountForUpdate :one
 SELECT * FROM accounts WHERE id = $1 FOR NO KEY UPDATE;
